consequently changed int32 to int since it brings no performance gain in any way

+++ src/backend/datastructures/graph.go
@@ -4,22 +4,22 @@ type Graph struct {
 	// holds the nodes of the Graph with its lat and long coordinates
 	Nodes [][]float64
 	// edge list that indicate to which node an edge points, its distance is saved in the 'offset' array
-	Edges []int32
+	Edges []int
 	// distance of and edge
-	Distance []int32
+	Distance []int
 	// for each node in nodes it holds the offset where to look for its outgoing edges in the 'edges' array
-	Offset []int32
+	Offset []int
 }
 
-func (g Graph) GetAllOutgoingEdgesOfNode(node int32) []int32 {
-	var outEdges []int32
+func (g Graph) GetAllOutgoingEdgesOfNode(node int) []int {
+	var outEdges []int
 	var nextOffset = 0
-	if node == int32(len(g.Nodes)-1) {
+	if node == (len(g.Nodes) - 1) {
 		nextOffset = len(g.Edges)
 	} else {
-		nextOffset = int(g.Offset[node+1])
+		nextOffset = g.Offset[node+1]
 	}
-	for i := g.Offset[node]; i < int32(nextOffset); i++ {
+	for i := g.Offset[node]; i < nextOffset; i++ {
 		outEdges = append(outEdges, i)
 	}
 	return outEdges
